pyproject.toml
[tool.poetry]
name = "auto"
version = "0.1.0"
description = ""
authors = ["Simon Butt <simonbutt123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.27.4"
google-cloud-aiplatform = "^1.23.0"
kfp = "^1.8.20"
langchain = "^0.0.139"
gradio = "^3.27.0"

[tool.poetry.group.dev.dependencies]
jupyterlab = "^3.6.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
